Use g_info and g_debug properly.
Create unknown processes with 0 creation time such as to reduce the risks of
thinking that two processes with the same pid are different because their
arbitrary creation time was different.

Prepare processTrace for the incremental background computation.

Use as key for processes in state.c the pid and cpu_name to account for 
the fact that there is an idle process (pid 0) for each cpu.



git-svn-id: http://ltt.polymtl.ca/svn@490 04897980-b3bd-0310-b5e0-8ef037075253