I recently checked out a later version of the viewer, and found a couple of build paths that cause problems with builds done outside the source directory.
Here's a patch:  (You can ignore the version numbers in the patch,
this applies OK against 0.8.79)

I patched both the Makefile.am and Makefile.in files, but obviously
if you re-gen the Makefile.in, that part of the patch is redundant.

The change is straightforward - it just uses top_builddir
in place of top_srcdir for the library path for linking.
When building within the source tree, top_builddir is the
same as top_srcdir, so this should have no effect for
people doing builds the "normal" way.

 lttv/modules/gui/tutorial/Makefile.am |    2 +-
 lttv/modules/text/Makefile.am         |    2 +-
 4 files changed, 4 insertions(+), 4 deletions(-)

<Signed-off-by: Tim Bird tim.bird (at) am.sony.com>



git-svn-id: http://ltt.polymtl.ca/svn@2474 04897980-b3bd-0310-b5e0-8ef037075253