Fix name table expend in state.c
Actually, looking back at state.c, I noticed there was a major problem
with name table resizing:

the top level pointers were copied from the name table structure to the
trace context, so when the tables were resized, the trace context was
not updated and still refered to the old tables. This is why we could
not free the old tables upon resizing.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>