Fix tracecontext seek backward underflow
Fix a problem with the detailed event list "seek backward". In the following
condition:

- Long interval between events (e.g. generated with power management suspend).
- Happening close to trace start.
- Trace start near 0s 0ns.

The substraction could underflow. Fix this by comparing the time to substract
and floor to trace start time if it would underflow.

The visible effect was that the detailed event list is seeked to the end of the
trace rather than the previous event when going "up" one event prior to the
suspend begin.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>
CC: Viktor Rosendahl <viktor.rosendahl@nokia.com>