Add formatted dump module v5
This new plugin prints a formated output of each events in a trace. The
output format is defined as a parameter with printf like syntax.It provides
a default format easy to read and the original textDump format for backward
compatibility.
Updated: Use glib functions for solving buffer overflow problem and correct
indentation. Add another format called "strace-like".

Signed-off-by: Vincent Attard <vincent.attard@polymtl.ca>
Reviewed-by: Yannick Brosseau <yannick.brosseau@polymtl.ca>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>